requirements.txt
# Core dependencies
openai==1.63.2
pdf2image==1.16.3
Pillow==9.5.0
python-dotenv==1.0.0
python-pptx==0.6.22
tqdm==4.66.1
requests>=2.31.0

# OpenAI dependencies
httpx>=0.23.0
pydantic>=1.9.0
typing-extensions>=4.11

# Required by python-pptx
XlsxWriter>=3.0.0
lxml>=4.9.0
